refactor: Transfer 클래스에 있던 메서드들 TransferTransaction 클래스로 위치 변
