+++ src/main/java/org/c4marathon/assignment/usecase/TransferUseCase.java
@@ -1,9 +1,10 @@
 package org.c4marathon.assignment.usecase;
 
-import org.c4marathon.assignment.domain.service.AccountPolicyService;
+import org.c4marathon.assignment.config.property.AccountPolicyProperties;
 import org.c4marathon.assignment.domain.service.mainaccount.MainAccountService;
-import org.c4marathon.assignment.domain.service.mainaccount.TransferRetryService;
+import org.c4marathon.assignment.domain.service.mainaccount.RetryService;
 import org.c4marathon.assignment.domain.validator.MainAccountValidator;
+import org.c4marathon.assignment.infra.retry.RetryExecutor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -15,10 +16,10 @@ import lombok.extern.slf4j.Slf4j;
 @RequiredArgsConstructor
 public class TransferUseCase {
 
+	private final AccountPolicyProperties accountPolicyProperties;
 	private final MainAccountValidator mainAccountValidator;
 	private final MainAccountService mainAccountService;
-	private final TransferRetryService transferRetryService;
-	private final AccountPolicyService accountPolicyService;
+	private final RetryService retryService;
 
 	// todo: 이체 실패 시, 충전 보상 로직 만들기
 	@Transactional
@@ -28,40 +29,13 @@ public class TransferUseCase {
 		Long shortfall = mainAccountService.calculateShortfall(fromAccountId, transferAmount);
 
 		if (shortfall <= 0) {
-			transferRetryService.transferWithRetry(fromAccountId, toAccountId, transferAmount);
+			retryService.transferWithRetry(fromAccountId, toAccountId, transferAmount);
 			return;
 		}
 
-		long chargeAmount = accountPolicyService.getRoundedCharge(shortfall);
+		long chargeAmount = accountPolicyProperties.getRoundedCharge(shortfall);
 		mainAccountService.chargeOrThrow(fromAccountId, chargeAmount, transferAmount);
 
-		transferRetryService.transferWithRetry(fromAccountId, toAccountId, transferAmount);
-	}
-
-	@Transactional
-	public void initiatePendingTransfer(Long fromAccountId, Long toAccountId, Long transferAmount) {
-		mainAccountValidator.validateTransfer(fromAccountId, toAccountId, transferAmount);
-
-		Long shortfall = mainAccountService.calculateShortfall(fromAccountId, transferAmount);
-
-		if (shortfall <= 0) {
-			transferRetryService.initiatePendingTransferWithRetry(fromAccountId, toAccountId, transferAmount);
-			return;
-		}
-
-		long chargeAmount = accountPolicyService.getRoundedCharge(shortfall);
-		mainAccountService.chargeOrThrow(fromAccountId, chargeAmount, transferAmount);
-
-		transferRetryService.initiatePendingTransferWithRetry(fromAccountId, toAccountId, transferAmount);
-	}
-
-	@Transactional
-	public void acceptPendingTransfer(Long transactionId) {
-		transferRetryService.acceptPendingTransferWthRetry(transactionId);
-	}
-
-	@Transactional
-	public void cancelPendingTransfer(Long transactionId) {
-		transferRetryService.cancelPendingTransferWithRetry(transactionId);
+		retryService.transferWithRetry(fromAccountId, toAccountId, transferAmount);
 	}
 }
